Cap index in navigateFragment
If a slide with fragments is the last in a vertical stack then navigateDown could go past the last fragment to a state where no fragments would be marked as current; this could also happen if the URL specifies a fragment index past the last one. This is most noticeable when highlight-current styles are used for the fragments. By limiting the index in nagvigateFragment to the largest possible index we prevent this occuring.